+++ src/api/utils.ts
@@ -401,3 +401,8 @@ export function convertToDataPoints(
     recurse(obj);
     return res;
 }
+
+export function isGuid(x: string): boolean {
+    const guidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+    return guidRegex.test(x);
+}