+++ src/cli/commands/mc-upload-file.ts
@@ -23,7 +23,7 @@ export default (program: CommanderStatic) => {
         .option("-f, --file <fileToUpload>", "file to upload to the file service")
         .option("-h, --filepath <filepath>", "file path in the mindsphere")
         .option("-l, --parallel <number>", "parallel chunk uploads", "3")
-        .option("-i, --assetid [assetid]", "asset id from the mindsphere  (default: upload to the agent)")
+        .option("-i, --assetid [assetid]", "mindsphere asset id  (default: upload to the agent)")
         .option("-m, --mime [mime-type]", "mime type of the file (default: automatic recognition)")
         .option("-d, --desc [description]", "description")
         .option("-k, --chunked", "Use chunked upload")