+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.12.0 - (Frost Vienna) - April 2021
+
+### Bugfixes and Improvements 3.12.0
+
+- TODO: bump AJV and nock
+- CLI: new `mc events` command for event management [#241]
+- CLI: renamed and extended `mc download-events` to `mc events-bulk` command with `--mode download` and  `--mode delete` [#240]
+- Bumped all dependencies
+
 ## 3.11.2 - (Sandcastle Vienna) - March 2021
 
 ### Bugfixes and Improvements 3.11.2